create main.py
TBH, this is a very basic script. I have not mastered python yet, so whatever elementary knowledge I had I used it to make the code. And yes, the code can be improved in various ways, but I don't have time since I am in Grade 12th in India. This code has many constraints, like to make it work on your computer you will need to mention the x, y coordinates of the respective buttons in all the click functions before running, tho it is a one time job.  Also, the option to use meeting ID to join a meeting can be done only on mac currently, since it uses 'spotlight'. Windows support might be added soon for the same once I get enough time. Although, this script can run on windows for joining through meeting link once the necessary one time changes in the x and y coordinates are made  according to the users display.